feat(apps/web): upgrade to react-router data router
