+++ apps/web/package.json
@@ -42,9 +42,8 @@
 		"prop-types": "^15.8.1",
 		"react": "18.2.0",
 		"react-dom": "18.2.0",
-		"react-icons": "^4.12.0",
-		"react-redux": "^8.1.1",
-		"react-router-dom": "^6.2.1",
+		"react-redux": "^9.1.2",
+		"react-router-dom": "^6.23.0",
 		"tailwindcss-animate": "^1.0.7",
 		"yup": "^1.2.0"
 	},
