+++ apps/web/src/scenes/invites/index.tsx
@@ -31,4 +31,4 @@ const Invites = () => {
 	);
 };
 
-export default Invites;
+export const Component = Invites;
