+++ apps/web/src/scenes/lists/index.tsx
@@ -1,13 +1,63 @@
-import React, { Suspense } from "react";
+import Button from "@hominem/components/Button";
+import { PlusCircle } from "lucide-react";
+import { useCallback, useState } from "react";
+import { useNavigate } from "react-router-dom";
 
-const LazyLists = React.lazy(
-	() => import(/* webpackChunkName: "lists" */ "./lists"),
-);
+import ListForm from "src/components/ListForm";
+import Lists from "src/components/Lists";
+import { useGetLists } from "src/lib/api";
+import { useAuth } from "src/lib/auth";
 
-const Dashboard = (props) => (
-	<Suspense>
-		<LazyLists {...props} />
-	</Suspense>
-);
+export function Component() {
+	const navigate = useNavigate();
+	const { user } = useAuth();
+	const [isListFormOpen, setIsListFormOpen] = useState(false);
+	const { data, error, refetch, status: listsStatus } = useGetLists();
 
-export default Dashboard;
+	const onAddListClick = useCallback(() => {
+		setIsListFormOpen(true);
+	}, []);
+
+	const onListCreate = useCallback(() => {
+		setIsListFormOpen(false);
+		refetch();
+	}, [refetch]);
+
+	const onListFormCancel = useCallback(() => {
+		setIsListFormOpen(false);
+	}, []);
+
+	if (!user) {
+		navigate("/");
+	}
+
+	return (
+		<div className="flex flex-col w-full">
+			<div className="flex items-center justify-between pb-4">
+				<h1 className="text-3xl font-bold">Lists</h1>
+				{!isListFormOpen ? (
+					<Button
+						className="bg-transparent"
+						onClick={onAddListClick}
+						disabled={isListFormOpen}
+					>
+						<PlusCircle size={24} className="text-primary" />
+					</Button>
+				) : null}
+			</div>
+			{isListFormOpen ? (
+				<div className="mb-4">
+					<ListForm onCancel={onListFormCancel} onCreate={onListCreate} />
+				</div>
+			) : null}
+			<div>
+				<Lists
+					status={listsStatus}
+					lists={data}
+					error={error}
+					currentUserEmail={user?.email}
+				/>
+			</div>
+		</div>
+	);
+}
