+++ apps/web/src/test/account.test.tsx
@@ -4,13 +4,12 @@ import { beforeEach, describe, expect, test } from "vitest";
 import { vi } from "vitest";
 
 import * as auth from "src/lib/auth";
+import { Component as Account } from "src/scenes/account";
 import {
 	TEST_USER_NAME,
 	renderWithProviders,
 	useAuthMock,
 } from "src/test/utils";
-
-import Account from "./Account";
 
 describe("Account", () => {
 	beforeEach(() => {
