+++ apps/web/src/test/place.test.tsx
@@ -2,10 +2,9 @@ import { screen } from "@testing-library/react";
 import { useParams } from "react-router-dom";
 import { type Mock, beforeEach, describe, expect, test } from "vitest";
 
+import { Component as Place } from "src/scenes/place";
 import { MOCK_PLACE } from "src/test/mocks/place";
 import { renderWithProviders } from "src/test/utils";
-
-import Place from "./Place";
 
 describe("Place", () => {
 	beforeEach(() => {
